+++ CHANGELOG.md
@@ -8,11 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## Version 4.0.28, 11/28/2024
+## Version 4.0.28, 11/29/2024
 
 ### Added
 
-Support for simple data type matching processing (text, substring, binary and b64)
+1. Support for simple data type matching processing (text, substring, binary and b64)
+2. Optional external state machine
 
 ### Removed
 
@@ -20,7 +21,7 @@ N/A
 
 ### Changed
 
-N/A
+Bugfix for AsyncHttpClient to allow missing HTTP request body in POST, PUT or PATCH request
 
 ---
 ## Version 4.0.27, 11/27/2024